@@ -9,7 +9,6 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
-import java.util.LinkedHashSet;
 import java.util.LinkedList;
 import java.util.Map;
 import java.util.Objects;
@@ -40,7 +39,6 @@ import graphql.language.FragmentSpread;
 import graphql.language.InlineFragment;
 import graphql.language.OperationDefinition;
 import graphql.language.SelectionSet;
-import lombok.val;
 
 public class OperationTranslator implements Translator {
 
@@ -79,7 +77,7 @@ public class OperationTranslator implements Translator {
 				Collection<FragmentDefinition> requiredFragments = new HashSet<>();
 				ctx.getDefinedSelections()
 					.computeIfAbsent(baseName, x -> traverseSelections(
-						GqlSelection.of(typeName, definition.getSelectionSet()), allFragments, requiredFragments, ctx));
+						typeName, definition.getSelectionSet(), allFragments, requiredFragments, ctx));
 				ctx.getDefinedOperations()
 					.computeIfAbsent(baseName, x -> GqlOperation.of(
 						definitionName,
@@ -96,7 +94,8 @@ public class OperationTranslator implements Translator {
 	}
 
 	private Map<String, Collection<GqlSelection>> traverseSelections(
-		GqlSelection rootSelection,
+		String typeName,
+		SelectionSet subset,
 		Collection<FragmentDefinition> allFragments,
 		Collection<FragmentDefinition> requiredFragments,
 		GqlContext ctx
@@ -104,23 +103,16 @@ public class OperationTranslator implements Translator {
 		Map<String, Map<Integer, Set<GqlSelection>>> typeMap = new HashMap<>();
 		Map<String, AtomicInteger> counters = new HashMap<>();
 		Queue<GqlSelection> selectionsToResolve = new LinkedList<>();
-		selectionsToResolve.offer(rootSelection);
+		selectionsToResolve.offer(GqlSelection.of(typeName, subset));
 		while (!selectionsToResolve.isEmpty()) {
-			val currentSelection = selectionsToResolve.poll();
-			String currentTypeName = currentSelection.getType().getInner();
+			var currentSelection = selectionsToResolve.poll();
+			String currentTypeName = currentSelection.getType().getName();
 			// get a set of selections by type name and their unresolved subset
-			Set<GqlSelection> subSelections = new HashSet<>();
-			Set<GqlSelection> unresolved = new LinkedHashSet<>();
-			currentSelection.getSubsets().forEach(unresolvedSet ->
-				resolveOneLevel(unresolvedSet, allFragments, requiredFragments, new HashSet<>(), ctx, currentTypeName)
-					.forEach(selection -> {
-						if (!selection.getSubsets().isEmpty()) {
-							unresolved.add(selection);
-						}
-						subSelections.add(selection);
-					}));
+			var subSelections = currentSelection.getSubsets().stream()
+				.flatMap(unresolved -> resolveOneLevel(unresolved, allFragments, requiredFragments, new HashSet<>(), ctx, currentTypeName).stream())
+				.collect(toSet());
 			// find exactly the same selection set already linked to this type
-			val variants = typeMap.computeIfAbsent(currentTypeName, x -> new HashMap<>());
+			var variants = typeMap.computeIfAbsent(currentTypeName, x -> new HashMap<>());
 			int variantNumber = variants.entrySet().stream()
 				.filter(entry -> entry.getValue().equals(subSelections))
 				.map(Map.Entry::getKey)
@@ -129,11 +121,11 @@ public class OperationTranslator implements Translator {
 					// this selection set has not been linked to this type before
 					int key = counters.computeIfAbsent(currentTypeName, x -> new AtomicInteger()).incrementAndGet();
 					variants.put(key, subSelections);
-					unresolved.forEach(selectionsToResolve::offer);
+					subSelections.stream().filter(s -> !s.getSubsets().isEmpty()).forEach(selectionsToResolve::offer);
 					return key;
 				});
 			// link previous selection to the type name
-			currentSelection.replaceTargetType(currentTypeName + getTypeSuffix(variantNumber));
+			currentSelection.setTargetTypeName(currentTypeName + getTypeSuffix(variantNumber));
 		}
 		Map<String, Collection<GqlSelection>> result = new HashMap<>();
 		typeMap.forEach((type, variants) -> variants.forEach((variantNumber, selections) ->
@@ -176,9 +168,11 @@ public class OperationTranslator implements Translator {
 			.filter(Optional::isPresent)
 			.map(Optional::get)
 			.filter(visitedFragments::add)
-			.peek(requiredFragments::add)
-			.map(fragment -> resolveOneLevel(fragment.getSelectionSet(),
-				allFragments, requiredFragments, visitedFragments, ctx, fragment.getTypeCondition().getName()))
+			.map(fragment -> {
+				requiredFragments.add(fragment);
+				return resolveOneLevel(fragment.getSelectionSet(),
+					allFragments, requiredFragments, visitedFragments, ctx, fragment.getTypeCondition().getName());
+			})
 			.flatMap(Collection::stream)
 			.forEach(selection -> result.merge(selection.getKey(), selection, GqlSelection::merge));
 		return result.values();
@@ -220,7 +214,7 @@ public class OperationTranslator implements Translator {
 			freemarker.getTemplate(OPERATION_DOCUMENT).process(null, writer);
 			return writer.toString();
 		} catch (TemplateException | IOException e) {
-			log.warn(String.format("Operation document [%s] is not created.", operation.getName()), e);
+			log.warn("Operation document [%s] is not created.".formatted(operation.getName()), e);
 			return null;
 		} finally {
 			freemarker.clearSharedVariables();
