@@ -16,24 +16,27 @@ public class ScalarMapParameterApplier implements ParameterApplier {
 
 	private static final String KEY_VALUE_SEPARATOR = "=";
 
+	private static final Map<String, String> DEFAULT_SCALARS = Map.of(
+		"Int", "Integer",
+		"Float", "Double",
+		"ID", "String"
+	);
+
 	@Override
 	public void apply(GqlConfigurationBuilder builder) {
-		builder
-			.scalar("Int", "Integer")
-			.scalar("Float", "Double")
-			.scalar("ID", "String");
+		builder.scalars(DEFAULT_SCALARS);
 		if (scalarMap != null) {
 			scalarMap.entrySet().stream()
 				.filter(item ->
 					item.getKey() != null && item.getValue() != null &&
-					item.getKey().trim().length() > 0 && item.getValue().trim().length() > 0)
+					!item.getKey().isBlank() && !item.getValue().isBlank())
 				.forEach(item -> builder.scalar(item.getKey().trim(), item.getValue().trim()));
 		}
 		else if (scalarMapAlternative != null) {
 			scalarMapAlternative.stream()
 				.filter(Objects::nonNull)
 				.map(item -> item.split(KEY_VALUE_SEPARATOR, 2))
-				.filter(item -> (item.length == 2) && item[0].trim().length() > 0 && item[1].trim().length() > 0)
+				.filter(item -> (item.length == 2) && !item[0].isBlank() && !item[1].isBlank())
 				.forEach(item -> builder.scalar(item[0].trim(), item[1].trim()));
 		}
 	}
